collision: switch to triangulated vertices
but getting some really weird issues. it works fine when the player X coordinate
is around 1500 and the Y coordinate is around 7000-8000, but otherwise exibits
some very strange behavior.

I looked at the code for a good two hours and couldn't make any sense of it. My
suspicion is that there was something wrong with the grid indexing, but that
doesn't explain why it worked with the poisson sample points.